requirements.txt
flask
torch
torchvision
transformers
pillow